requirements.txt
asyncio
langchain>=0.1.6
langchain-experimental>=0.0.50
openai>=1.12.0
ragas>=0.1.0
chromadb>=0.4.22
python-dotenv
pypdf